update pkg to dart 3.9.2

+++ inner/pkg/compiler/test/codegen/type_inference2_test.dart
@@ -2,13 +2,13 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:async_helper/async_helper.dart';
+import 'package:expect/async_helper.dart';
 import '../helpers/compiler_helper.dart';
 
 const String TEST_ONE = r"""
 sum(param0, param1) {
   var sum = 0;
-  for (var i = param0; i < param1; i += 1) sum = sum + i;
+  for (var i = param0; i < param1; i += 1) sum = sum + i as int;
   return sum;
 }
 """;
